feat: replace DayView with unified Timeline component
- Delete DayView directory (merged into Timeline)
- Port DayView CSS to timeline-* class names
- Remove /day route and Day nav link
- Update AddData to navigate to /timeline after creating data
- Update ScreentimeCategories to invalidate timeline-productivity query key
- Add activityMerge.ts with pure functions for unified Activity column
- Add tests for activityMerge.ts (22 tests all passing)
- Refactor buildActivityColumnItems to extract helpers reducing complexity
