--- apps/web/src/pages/DayView/style.css
@@ -1,393 +1,0 @@
-.day-view {
-  margin: 0 auto;
-  padding: 2rem;
-  width: 100%;
-}
-
-.day-view h1 {
-  font-size: 2rem;
-  margin: 0 0 1.5rem 0;
-  color: #1f2937;
-}
-
-/* Controls */
-.day-view-controls {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 1rem;
-  margin-bottom: 1.5rem;
-  padding: 0.75rem 1rem;
-  background: #f9fafb;
-  border-radius: 8px;
-  border: 1px solid #e5e7eb;
-}
-
-/* Navigation buttons */
-.day-view-nav {
-  display: flex;
-  align-items: center;
-  gap: 0.25rem;
-}
-
-.day-view-nav .nav-btn {
-  background: transparent;
-  border: 1px solid #d1d5db;
-  border-radius: 4px;
-  padding: 0.375rem 0.625rem;
-  font-size: 0.875rem;
-  font-weight: 600;
-  cursor: pointer;
-  color: #374151;
-  transition:
-    background 0.15s,
-    border-color 0.15s;
-  min-width: 2rem;
-  text-align: center;
-}
-
-.day-view-nav .nav-btn:hover {
-  background: #e5e7eb;
-  border-color: #9ca3af;
-}
-
-.day-view-nav .nav-btn:active {
-  background: #d1d5db;
-}
-
-.day-view-nav .nav-today {
-  padding: 0.375rem 0.75rem;
-  font-weight: 500;
-}
-
-/* Date label */
-.day-view-date-label {
-  font-size: 0.9rem;
-  color: #374151;
-  font-weight: 500;
-}
-
-/* Fetching indicator */
-.day-view-fetching {
-  font-size: 0.8rem;
-  color: #6b7280;
-  animation: dayview-pulse 1.5s ease-in-out infinite;
-}
-
-@keyframes dayview-pulse {
-  0%,
-  100% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.4;
-  }
-}
-
-/* Legend */
-.day-view-legend {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.75rem;
-  padding: 0.75rem 1rem;
-  margin-bottom: 1rem;
-  background: #f9fafb;
-  border-radius: 8px;
-  border: 1px solid #e5e7eb;
-  font-size: 0.85rem;
-}
-
-.day-view-legend .legend-item {
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  cursor: pointer;
-  background: none;
-  border: none;
-  padding: 0.25rem 0.5rem;
-  border-radius: 4px;
-  font: inherit;
-  font-size: 0.85rem;
-  color: inherit;
-  transition:
-    opacity 0.15s,
-    background 0.15s;
-}
-
-.day-view-legend .legend-item:hover {
-  background: rgba(0, 0, 0, 0.05);
-}
-
-.day-view-legend .legend-item-hidden {
-  opacity: 0.4;
-  text-decoration: line-through;
-}
-
-.day-view-legend .legend-dot {
-  width: 12px;
-  height: 12px;
-  border-radius: 3px;
-  flex-shrink: 0;
-}
-
-.day-view-legend .legend-separator {
-  width: 1px;
-  height: 20px;
-  background: #d1d5db;
-  margin: 0 0.25rem;
-  flex-shrink: 0;
-}
-
-/* Chart container */
-.day-view-chart-container {
-  position: relative;
-  overflow: hidden;
-}
-
-.day-view-chart-container svg {
-  display: block;
-  cursor: grab;
-}
-
-.day-view-chart-container svg:active {
-  cursor: grabbing;
-}
-
-.day-view-chart-container svg [data-clickable] {
-  cursor: pointer;
-}
-
-/* Tooltip */
-.day-view-tooltip {
-  position: absolute;
-  pointer-events: none;
-  background: rgba(0, 0, 0, 0.9);
-  color: white;
-  padding: 0.625rem 0.875rem;
-  border-radius: 6px;
-  font-size: 0.8rem;
-  line-height: 1.5;
-  max-width: 300px;
-  z-index: 1000;
-  white-space: pre-line;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
-}
-
-.day-view-tooltip .tooltip-title {
-  font-weight: 600;
-  margin-bottom: 0.25rem;
-}
-
-.day-view-tooltip .tooltip-time {
-  color: #9ca3af;
-  font-size: 0.75rem;
-}
-
-.day-view-tooltip .tooltip-detail {
-  color: #d1d5db;
-  margin-top: 0.25rem;
-}
-
-.day-view-tooltip .tooltip-music {
-  color: #ec4899;
-  margin-top: 0.375rem;
-  font-style: italic;
-}
-
-.day-view-tooltip .hr-zone-bar {
-  display: flex;
-  height: 8px;
-  border-radius: 4px;
-  overflow: hidden;
-  margin-top: 0.375rem;
-}
-
-.day-view-tooltip .hr-zone-bar span {
-  display: block;
-  height: 100%;
-}
-
-/* Help text */
-.day-view-help {
-  font-size: 0.8rem;
-  color: #9ca3af;
-  margin-top: 0.5rem;
-  text-align: center;
-}
-
-/* Loading and error states */
-.day-view .loading,
-.day-view .error {
-  text-align: center;
-  padding: 2rem;
-  color: #6b7280;
-}
-
-.day-view .error {
-  color: #dc2626;
-}
-
-/* Mobile list fallback */
-.day-view-list {
-  display: none;
-}
-
-.day-view-list-item {
-  display: flex;
-  align-items: flex-start;
-  gap: 0.75rem;
-  padding: 0.75rem 0;
-  border-bottom: 1px solid #e5e7eb;
-  text-decoration: none;
-  color: inherit;
-}
-
-.day-view-list-item.clickable {
-  cursor: pointer;
-}
-
-.day-view-list-item.clickable:hover {
-  background: #f9fafb;
-}
-
-.day-view-list-item .list-dot {
-  width: 10px;
-  height: 10px;
-  border-radius: 3px;
-  flex-shrink: 0;
-  margin-top: 0.375rem;
-}
-
-.day-view-list-item .list-time {
-  font-size: 0.8rem;
-  color: #6b7280;
-  min-width: 90px;
-  flex-shrink: 0;
-}
-
-.day-view-list-item .list-content {
-  flex: 1;
-}
-
-.day-view-list-item .list-title {
-  font-weight: 500;
-  color: #374151;
-}
-
-.day-view-list-item .list-detail {
-  font-size: 0.85rem;
-  color: #6b7280;
-}
-
-/* Column headers */
-.day-view-column-headers {
-  display: flex;
-  font-size: 0.8rem;
-  font-weight: 600;
-  color: #6b7280;
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-  margin-bottom: 0.25rem;
-  padding: 0 0 0.5rem 0;
-  border-bottom: 1px solid #e5e7eb;
-}
-
-/* Dark mode */
-@media (prefers-color-scheme: dark) {
-  .day-view h1 {
-    color: #f3f4f6;
-  }
-
-  .day-view-controls {
-    background: #1f2937;
-    border-color: #374151;
-  }
-
-  .day-view-nav .nav-btn {
-    color: #d1d5db;
-    border-color: #4b5563;
-    background: transparent;
-  }
-
-  .day-view-nav .nav-btn:hover {
-    background: #374151;
-    border-color: #6b7280;
-  }
-
-  .day-view-nav .nav-btn:active {
-    background: #4b5563;
-  }
-
-  .day-view-date-label {
-    color: #d1d5db;
-  }
-
-  .day-view-fetching {
-    color: #9ca3af;
-  }
-
-  .day-view-legend {
-    background: #1f2937;
-    border-color: #374151;
-  }
-
-  .day-view-legend .legend-item:hover {
-    background: rgba(255, 255, 255, 0.08);
-  }
-
-  .day-view-legend .legend-separator {
-    background: #4b5563;
-  }
-
-  .day-view-tooltip {
-    background: rgba(17, 24, 39, 0.95);
-  }
-
-  .day-view-column-headers {
-    color: #9ca3af;
-    border-color: #374151;
-  }
-
-  .day-view-list-item {
-    border-color: #374151;
-  }
-
-  .day-view-list-item .list-title {
-    color: #d1d5db;
-  }
-
-  .day-view-list-item .list-detail {
-    color: #9ca3af;
-  }
-
-  .day-view-list-item.clickable:hover {
-    background: #1f2937;
-  }
-}
-
-/* Responsive */
-@media (max-width: 768px) {
-  .day-view {
-    padding: 1rem;
-  }
-
-  .day-view h1 {
-    font-size: 1.5rem;
-  }
-
-  .day-view-chart-container {
-    display: none;
-  }
-
-  .day-view-column-headers {
-    display: none;
-  }
-
-  .day-view-help {
-    display: none;
-  }
-
-  .day-view-list {
-    display: block;
-  }
-}
