+++ apps/web/src/pages/Timeline/drawActivitySparklines.ts
