+++ apps/web/src/pages/Timeline/style.css
@@ -1,13 +1,12 @@
-.timeline {
+.timeline-view {
+  margin: 0 auto;
+  padding: 2rem;
   width: 100%;
-  margin: 0;
-  padding: 1rem 1.5rem;
-  box-sizing: border-box;
 }
 
-.timeline h1 {
+.timeline-view h1 {
   font-size: 2rem;
-  margin: 0 0 1rem 0;
+  margin: 0 0 1.5rem 0;
   color: #1f2937;
 }
 
@@ -17,7 +16,7 @@
   flex-wrap: wrap;
   align-items: center;
   gap: 1rem;
-  margin-bottom: 1rem;
+  margin-bottom: 1.5rem;
   padding: 0.75rem 1rem;
   background: #f9fafb;
   border-radius: 8px;
@@ -61,14 +60,35 @@
   font-weight: 500;
 }
 
+/* Date label */
+.timeline-date-label {
+  font-size: 0.9rem;
+  color: #374151;
+  font-weight: 500;
+}
+
+/* Orientation toggle */
+.timeline-orientation-toggle {
+  display: flex;
+  align-items: center;
+  gap: 0.25rem;
+  margin-left: auto;
+}
+
+.timeline-orientation-toggle .nav-btn.active {
+  background: #374151;
+  color: #fff;
+  border-color: #374151;
+}
+
 /* Fetching indicator */
 .timeline-fetching {
   font-size: 0.8rem;
   color: #6b7280;
-  animation: pulse 1.5s ease-in-out infinite;
+  animation: timeline-pulse 1.5s ease-in-out infinite;
 }
 
-@keyframes pulse {
+@keyframes timeline-pulse {
   0%,
   100% {
     opacity: 1;
@@ -78,29 +98,7 @@
   }
 }
 
-/* Layer toggles */
-.timeline-layers {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-  margin-bottom: 1rem;
-}
-
-.timeline-layers label {
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  font-size: 0.9rem;
-  cursor: pointer;
-}
-
-.timeline-layers input[type='checkbox'] {
-  width: 1rem;
-  height: 1rem;
-  cursor: pointer;
-}
-
-/* Legends */
+/* Legend */
 .timeline-legend {
   display: flex;
   flex-wrap: wrap;
@@ -113,20 +111,44 @@
   font-size: 0.85rem;
 }
 
-.timeline-legend strong {
-  color: #374151;
-}
-
 .timeline-legend .legend-item {
   display: flex;
   align-items: center;
   gap: 0.375rem;
+  cursor: pointer;
+  background: none;
+  border: none;
+  padding: 0.25rem 0.5rem;
+  border-radius: 4px;
+  font: inherit;
+  font-size: 0.85rem;
+  color: inherit;
+  transition:
+    opacity 0.15s,
+    background 0.15s;
+}
+
+.timeline-legend .legend-item:hover {
+  background: rgba(0, 0, 0, 0.05);
+}
+
+.timeline-legend .legend-item-hidden {
+  opacity: 0.4;
+  text-decoration: line-through;
 }
 
 .timeline-legend .legend-dot {
   width: 12px;
   height: 12px;
-  border-radius: 50%;
+  border-radius: 3px;
+  flex-shrink: 0;
+}
+
+.timeline-legend .legend-separator {
+  width: 1px;
+  height: 20px;
+  background: #d1d5db;
+  margin: 0 0.25rem;
   flex-shrink: 0;
 }
 
@@ -167,32 +189,39 @@
   font-style: italic;
 }
 
-/* Chart container - full width */
+/* Chart container */
 .timeline-chart-container {
   position: relative;
-  width: 100%;
   overflow: hidden;
 }
 
 .timeline-chart-container svg {
   display: block;
-  width: 100%;
+  cursor: grab;
+}
+
+.timeline-chart-container svg:active {
+  cursor: grabbing;
+}
+
+.timeline-chart-container svg [data-clickable] {
+  cursor: pointer;
 }
 
 /* Tooltip */
 .timeline-tooltip {
   position: absolute;
   pointer-events: none;
-  background: rgba(0, 0, 0, 0.85);
+  background: rgba(0, 0, 0, 0.9);
   color: white;
-  padding: 0.5rem 0.75rem;
+  padding: 0.625rem 0.875rem;
   border-radius: 6px;
   font-size: 0.8rem;
-  line-height: 1.4;
+  line-height: 1.5;
   max-width: 300px;
   z-index: 1000;
-  white-space: nowrap;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+  white-space: pre-line;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
 }
 
 .timeline-tooltip .tooltip-title {
@@ -205,28 +234,28 @@
   font-size: 0.75rem;
 }
 
-.timeline-tooltip .tooltip-value {
-  color: #60a5fa;
-}
-
-.timeline-tooltip .tooltip-warning {
-  color: #fbbf24;
-  font-size: 0.75rem;
+.timeline-tooltip .tooltip-detail {
+  color: #d1d5db;
   margin-top: 0.25rem;
-  border-top: 1px solid rgba(251, 191, 36, 0.3);
-  padding-top: 0.25rem;
 }
 
-/* Loading and error states */
-.timeline .loading,
-.timeline .error {
-  text-align: center;
-  padding: 2rem;
-  color: #6b7280;
+.timeline-tooltip .tooltip-music {
+  color: #ec4899;
+  margin-top: 0.375rem;
+  font-style: italic;
 }
 
-.timeline .error {
-  color: #dc2626;
+.timeline-tooltip .hr-zone-bar {
+  display: flex;
+  height: 8px;
+  border-radius: 4px;
+  overflow: hidden;
+  margin-top: 0.375rem;
+}
+
+.timeline-tooltip .hr-zone-bar span {
+  display: block;
+  height: 100%;
 }
 
 /* Help text */
@@ -237,20 +266,92 @@
   text-align: center;
 }
 
+/* Loading and error states */
+.timeline-view .loading,
+.timeline-view .error {
+  text-align: center;
+  padding: 2rem;
+  color: #6b7280;
+}
+
+.timeline-view .error {
+  color: #dc2626;
+}
+
+/* Mobile list fallback */
+.timeline-list {
+  display: none;
+}
+
+.timeline-list-item {
+  display: flex;
+  align-items: flex-start;
+  gap: 0.75rem;
+  padding: 0.75rem 0;
+  border-bottom: 1px solid #e5e7eb;
+  text-decoration: none;
+  color: inherit;
+}
+
+.timeline-list-item.clickable {
+  cursor: pointer;
+}
+
+.timeline-list-item.clickable:hover {
+  background: #f9fafb;
+}
+
+.timeline-list-item .list-dot {
+  width: 10px;
+  height: 10px;
+  border-radius: 3px;
+  flex-shrink: 0;
+  margin-top: 0.375rem;
+}
+
+.timeline-list-item .list-time {
+  font-size: 0.8rem;
+  color: #6b7280;
+  min-width: 90px;
+  flex-shrink: 0;
+}
+
+.timeline-list-item .list-content {
+  flex: 1;
+}
+
+.timeline-list-item .list-title {
+  font-weight: 500;
+  color: #374151;
+}
+
+.timeline-list-item .list-detail {
+  font-size: 0.85rem;
+  color: #6b7280;
+}
+
+/* Column headers */
+.timeline-column-headers {
+  display: flex;
+  font-size: 0.8rem;
+  font-weight: 600;
+  color: #6b7280;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+  margin-bottom: 0.25rem;
+  padding: 0 0 0.5rem 0;
+  border-bottom: 1px solid #e5e7eb;
+}
+
 /* Dark mode */
 @media (prefers-color-scheme: dark) {
-  .timeline h1 {
+  .timeline-view h1 {
     color: #f3f4f6;
   }
 
-  .timeline-controls,
-  .timeline-legend {
+  .timeline-controls {
     background: #1f2937;
     border-color: #374151;
-  }
-
-  .timeline-controls label {
-    color: #d1d5db;
   }
 
   .timeline-nav .nav-btn {
@@ -268,16 +369,56 @@
     background: #4b5563;
   }
 
+  .timeline-orientation-toggle .nav-btn.active {
+    background: #d1d5db;
+    color: #111827;
+    border-color: #d1d5db;
+  }
+
+  .timeline-date-label {
+    color: #d1d5db;
+  }
+
   .timeline-fetching {
     color: #9ca3af;
   }
 
-  .timeline-legend strong {
-    color: #d1d5db;
+  .timeline-legend {
+    background: #1f2937;
+    border-color: #374151;
+  }
+
+  .timeline-legend .legend-item:hover {
+    background: rgba(255, 255, 255, 0.08);
+  }
+
+  .timeline-legend .legend-separator {
+    background: #4b5563;
   }
 
   .timeline-tooltip {
     background: rgba(17, 24, 39, 0.95);
+  }
+
+  .timeline-column-headers {
+    color: #9ca3af;
+    border-color: #374151;
+  }
+
+  .timeline-list-item {
+    border-color: #374151;
+  }
+
+  .timeline-list-item .list-title {
+    color: #d1d5db;
+  }
+
+  .timeline-list-item .list-detail {
+    color: #9ca3af;
+  }
+
+  .timeline-list-item.clickable:hover {
+    background: #1f2937;
   }
 
   .timeline-overlap-warnings {
@@ -297,16 +438,27 @@
 
 /* Responsive */
 @media (max-width: 768px) {
-  .timeline {
-    padding: 0.75rem;
+  .timeline-view {
+    padding: 1rem;
   }
 
-  .timeline h1 {
+  .timeline-view h1 {
     font-size: 1.5rem;
   }
 
-  .timeline-controls {
-    flex-direction: column;
-    align-items: flex-start;
+  .timeline-chart-container {
+    display: none;
+  }
+
+  .timeline-column-headers {
+    display: none;
+  }
+
+  .timeline-help {
+    display: none;
+  }
+
+  .timeline-list {
+    display: block;
   }
 }
