+++ apps/web/src/pages/Timeline/types.ts
@@ -1,4 +1,13 @@
-export type Column = 'Sleep / Rest' | 'Exercise' | 'Location' | 'Tags / Events' | 'Screen Time' | 'Music'
+export type Column =
+  | 'Activity'
+  | 'Sleep / Rest'
+  | 'Exercise'
+  | 'Location'
+  | 'Tags / Events'
+  | 'Screen Time'
+  | 'Music'
+
+export type Orientation = 'horizontal' | 'vertical'
 
 export interface TooltipContent {
   title: string
